RL_Foundation_Mujoco_including_DT_MGDT_GATO_and_TIT/requirements.txt
absl-py==1.4.0
ale-py==0.8.1
astunparse==1.6.3
atari-py==0.2.9
blosc==1.11.1
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chex==0.1.6
click==8.1.3
cloudpickle==2.2.1
configparser==5.3.0
contourpy==1.0.7
cycler==0.11.0
cython==0.29.33
d4rl==1.1
decorator==5.1.1
dm-control==1.0.10
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
dopamine-rl==4.0.6
etils==1.0.0
fasteners==0.18
filelock==3.9.0
flatbuffers==23.1.21
flax==0.6.6
fonttools==4.38.0
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.10
gitpython==3.1.31
glfw==2.5.6
google-auth-oauthlib==0.4.6
google-auth==2.16.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.51.3
gym-notices==0.0.8
gym==0.23.1
h5py==3.8.0
huggingface-hub==0.12.1
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
jax==0.4.4
jaxlib==0.4.4
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
labmaze==1.0.6
libclang==15.0.6.1
lxml==4.9.2
markdown-it-py==2.2.0
markdown==3.4.1
markupsafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
mj-envs==1.0.0
mjrl==1.0.0
msgpack==1.0.4
mujoco-py==2.1.2.14
mujoco==2.3.2
mypy-extensions==1.0.0
numpy==1.24.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax==0.1.4
orbax==0.1.2
packaging==23.0
pandas==1.5.3
pillow==9.4.0
pip==22.3.1
promise==2.3
protobuf==3.19.6
psutil==5.9.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pybullet==3.2.5
pycparser==2.21
pygame==2.2.0
pygments==2.14.0
pyopengl==3.1.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pyyaml==6.0
regex==2022.10.31
requests-oauthlib==1.3.1
requests==2.28.2
rich==13.3.1
rsa==4.9
sacremoses==0.0.53
scikit-video==1.1.11
scipy==1.10.1
sentry-sdk==1.15.0
setuptools==65.6.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
style==1.1.0
subprocess32==3.5.4
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.2
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-probability==0.19.0
tensorflow==2.11.0
tensorstore==0.1.33
termcolor==2.2.0
tf-slim==1.1.0
tokenizers==0.10.3
toolz==0.12.0
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm==4.64.1
transformers==4.11.3
typed-argument-parser==1.7.2
typing-extensions==4.5.0
typing-inspect==0.8.0
update==0.0.1
urllib3==1.26.14
wandb==0.9.1
watchdog==2.3.0
werkzeug==2.2.3
wheel==0.38.4
wrapt==1.14.1
zipp==3.14.0